google meet to group and trello ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
pydantic==2.5.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
py-trello==0.19.0
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3
openai==1.3.7
httpx==0.25.2
flask==3.0.0
flask-cors==4.0.0